chore: move matchers test out of examples
